Zabbix: Added Servers Status
General: Arranged files
