+++ lib/core/models/zabbix_api/zabbix_interface.dart
@@ -1,0 +1,17 @@
+/// This is used to store interface data from a host.
+class ZabbixInterface {
+  final String dns;
+  final String ip;
+
+  const ZabbixInterface({
+    this.dns,
+    this.ip,
+  });
+
+  factory ZabbixInterface.fromJson(Map<String, dynamic> json) {
+    return ZabbixInterface(
+      dns: json['dns'],
+      ip: json['ip'],
+    );
+  }
+}
