+++ lib/ui/widgets/zabbix_widgets/zabbix_widgets.dart
@@ -1,0 +1,3 @@
+export 'zabbix_ap.dart';
+export 'zabbix_servers.dart';
+export 'zabbix_switches.dart';
